+++ devel/tinytest/test-strrep.R
@@ -14,8 +14,8 @@ for (args in list(
     list(1:10, integer(0)),
     list(c(1, NA, 3, 5), c(NA, 3))
 )) {
-    expect_equivalent(do.call(stringx::`%x*%`, args), do.call(base::strrep, args))
-    expect_equivalent(do.call(stringx::strrep, args), do.call(base::strrep, args))
+    expect_equal(do.call(stringx::`%x*%`, args), do.call(base::strrep, args))
+    expect_equal(do.call(stringx::strrep, args), do.call(base::strrep, args))
 }
 
 
@@ -27,14 +27,20 @@ rm(`*.character`)
 
 
 
-x <- structure(c(A=1, B=NA, C=3), attrib1="value1")
-expect_equivalent(x %x*% 3, base::strrep(x, 3))
-expect_equivalent(stringx::strrep(x, 3), base::strrep(x, 3))
-expect_equivalent(sorted_attributes(x %x*% 1:2), sorted_attributes(x * 1:2))
+x <- structure(c(A="a", B=NA, C="b"), attrib1="value1")
+expect_equal(as.character(x %x*% 3), as.character(base::strrep(x, 3)))
+expect_equal(as.character(stringx::strrep(x, 3)), as.character(base::strrep(x, 3)))
+expect_equal(sorted_attributes(x %x*% 3), sorted_attributes(x))
+expect_equal(sorted_attributes(stringx::strrep(x, 3)), sorted_attributes(x))
+
+# only names.... and only if input is of type character
+expect_equal(sorted_attributes(base::strrep(x, 3)), list(names=names(x)))
+
+x <- matrix(letters[1:6], nrow=2, dimnames=list(c("A", "B"), NULL))
+expect_equal(stringx::strrep(x, 1:2), x %x*% 1:2)
 
 x <- matrix(1:6, nrow=2, dimnames=list(c("A", "B"), NULL))
-expect_equivalent(sorted_attributes(x %x*% 1:2), sorted_attributes(x * 1:2))
-expect_equivalent(stringx::strrep(x, 1:2), base::strrep(x, 1:2))
+expect_equal(sorted_attributes(x %x*% 1:2), sorted_attributes(x * 1:2))
 
 expect_warning(stringx::strrep(1:3, 1:2))
 expect_silent(base::strrep(1:3, 1:2))  # inconsistent